# internal/docs/swagger.yaml
info:
  contact: {}
paths:
  /v1/interactions:
    post:
      consumes:
      - application/json
      description: Create new interaction
      parameters:
      - description: User ID
        in: path
        name: user_id
        required: true
        type: string
      - description: Post ID
        in: path
        name: post_id
        required: true
        type: string
      - description: Interaction Type
        in: path
        name: interaction_type
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "500":
          description: Internal Server Error
      security:
      - ApiKeyAuth: []
      summary: Create new interaction
      tags:
      - interaction
  /v1/rankings:
    get:
      consumes:
      - application/json
      description: Get global ranking
      parameters:
      - description: Limit
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              type: string
            type: array
        "400":
          description: Bad Request
        "500":
          description: Internal Server Error
      summary: Get global ranking
      tags:
      - rankings
  /v1/rankings/{user_id}:
    get:
      consumes:
      - application/json
      description: Get personal ranking
      parameters:
      - description: User ID
        in: path
        name: user_id
        required: true
        type: string
      - description: Limit
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              type: string
            type: array
        "400":
          description: Bad Request
        "500":
          description: Internal Server Error
      summary: Get personal ranking
      tags:
      - rankings
swagger: "2.0"